Refactor and move content quota check to routers.py from models.py
